user.java return statement in getId changed

+++ src/main/java/com/emmabanuelos/ParentPortal/models/User.java
@@ -34,6 +34,6 @@ public class User extends AbstractEntity {
 
 
     public int getId() {
-    return (int) id;
+    return id;
     }
 }
